+++ common_func.py
@@ -4,51 +4,28 @@ Common functions used in flux calculation
 (c) Wu Sun <wu.sun@ucla.edu> 2016-2017
 
 """
+from collections import namedtuple
+import warnings
+
 import numpy as np
-import pandas as pd
 from scipy import optimize
 import scipy.constants.constants as sci_const
-# import datetime
-import warnings
-from collections import namedtuple
+import pandas as pd
 import yaml
 
 
-T_0 = sci_const.zero_Celsius
-
-
-def resistant_mean(x, IQR_range=1.5):
-    """
-    Input: 1D `numpy.ndarray` like
-    Output: mean with outliers removed
-    """
-    x = np.array(x)
-    if np.sum(np.isfinite(x)) <= 1:
-        return(np.nanmean(x))
-    else:
-        x_q1, x_q3 = np.nanpercentile(x, [25, 75])
-        x_iqr = x_q3 - x_q1
-        x_uplim = x_q3 + IQR_range * x_iqr
-        x_lolim = x_q1 - IQR_range * x_iqr
-        x_rmean = np.nanmean(x[(x >= x_lolim) & (x <= x_uplim)])
-        return(x_rmean)
-
-
-def resistant_std(x, IQR_range=1.5):
-    """
-    Input: 1D `numpy.ndarray` like
-    Output: standard deviation (deg. freedom = 1) with outliers removed
-    """
-    x = np.array(x)
-    if np.sum(np.isfinite(x)) <= 1:
-        return(np.nanstd(x, ddof=1))
-    else:
-        x_q1, x_q3 = np.nanpercentile(x, [25, 75])
-        x_iqr = x_q3 - x_q1
-        x_uplim = x_q3 + IQR_range * x_iqr
-        x_lolim = x_q1 - IQR_range * x_iqr
-        x_rstd = np.nanstd(x[(x >= x_lolim) & (x <= x_uplim)], ddof=1)
-        return(x_rstd)
+# Physical constants
+# Do not modify unless you are on a differnt planet or in a different universe.
+# - 'p_std': 1 standard atmospheric pressure in Pascal
+# - 'R_gas': The universal gas constant in J mol^-1 K^-1
+# - 'T_0': zero Celsius in Kelvin
+# - 'air_conc_std': Air concentration (mol m^-3) at STP condition
+phys_const = {
+    'T_0': sci_const.zero_Celsius,
+    'p_std': sci_const.atm,
+    'R_gas': sci_const.R,
+    'air_conc_std': sci_const.atm / (sci_const.R * sci_const.zero_Celsius), }
+T_0 = phys_const['T_0']
 
 
 def chamber_lookup_table_func(doy, chamber_config=None):
@@ -67,7 +44,8 @@ def chamber_lookup_table_func(doy, chamber_config=None):
             chamber_config = yaml.load(fo)
 
     for key in chamber_config:
-        if chamber_config[key]['schedule_start'] <= doy < chamber_config[key]['schedule_end']:
+        if (chamber_config[key]['schedule_start'] <= doy <
+                chamber_config[key]['schedule_end']):
             current_schedule = chamber_config[key]
             break
     else:
@@ -83,8 +61,9 @@ def chamber_lookup_table_func(doy, chamber_config=None):
 
     df = pd.DataFrame()
 
-    for key in ['ch_no', 'A_ch', 'A_ch_std', 'V_ch', 'ch_label', 'flowmeter_no',
-                'TC_no', 'PAR_no', 'ch_start', 'ch_o_b', 'ch_cls', 'ch_o_a',
+    for key in ['ch_no', 'A_ch', 'A_ch_std', 'V_ch', 'ch_label',
+                'flowmeter_no', 'TC_no', 'PAR_no',
+                'ch_start', 'ch_o_b', 'ch_cls', 'ch_o_a',
                 'ch_end', 'ch_atm_a']:
         df[key] = current_schedule[key]
 
@@ -113,258 +92,166 @@ def chamber_lookup_table_func(doy, chamber_config=None):
     return chamber_lookup_table
 
 
-def chamber_lookup_table_func_old(doy, return_all=False):
-    """
-    Return a chamber meta information look-up table (pandas.DataFrame). 
+def timelag_optmz_func():
+    # @TODO: to add this function
+    return None
 
-    Parameters
-    ----------
-    doy : float
-        Days since Jan 1 00:00 of the current year.
-    return_all : boolean, optional
-        By default `False`, which means only the chamber lookup table is returned.
-        If set as `True`, the function also returns other related information (see "Returns"). 
 
-    Returns
-    -------
-    chlut : pandas.DataFrame
-        Chamber meta information lookup table, including these columns:
-        * 'ch_no': the sequence of chamber numbers in a full cycle
-        * 'ch_label': strings used to label the chambers
-        * 'flowmeter_no': ID numbers of flowmeters corresponding to the chambers; assign -1 if not existent
-        * 'TC_no': ID numbers of temperature readings corresponding to the chamber numbers; assign -1 if not existent
-        * 'PAR_no': ID numbers of PAR readings corresponding to the chamber numbers; assign -1 if not existent
-        * 'A_ch': chamber-covered surface areas (for soil flux) or leaf areas (for leaf flux)
-        * 'A_ch_std': standard deviations of chamber-covered surface areas (for soil flux) or leaf areas (for leaf flux)
-        * 'V_ch': chamber volumes
-        * 'ch_start': starting times of chamber sampling with respect to the start of each cycle, in fraction of a day
-        * 'ch_o_b': times when chambers open, with respect to 'ch_start', in fraction of a day
-        * 'ch_cls': times when chambers close, with respect to 'ch_start', in fraction of a day
-        * 'ch_o_a': times when chambers reopen after a closure period, with respect to 'ch_start', in fraction of a day
-        * 'ch_end': times when chamber sampling ends, with respect to 'ch_start', in fraction of a day
-    Note: you may change the intermediate steps that generate `chlut` as you like, since they are only used for 
-        getting the lookup table. As long as the returned `chlut` has the above columns properly assigned, it should 
-        work well with the main program (`flux_calc.py`). 
+def volume_eff_optmz_func():
+    # @TODO: to add this function
+    return None
 
-    If `return_all == True`, return also these variables:
-    `n_ch` : integer
-        The number of chambers being alternated in this schedule. 
-        Note: a chamber may be sampled more than once in a full cycle, for example, chamber sequence 1-2-3-3-2-1. 
-    `smpl_cycle_len` : float
-        The duration of a full cycle over all chambers, in fraction of a day
-    `n_cycle_per_day` : integer
-        The number of full sampling cycles in a day. 
-    `next_schedule_switch` : float
-        The time (in day of year) when the sampling schedule switched to the next one. 
-    
-    Raises
-    ------
-    None
-    """
-    # chamber parameters
-    # there are 6 unique chambers in this example
-    # to be assigned: chamber volumes and surface (leaf) areas, `V_ch` and `A_ch`
-    V_ch = np.zeros(6) * np.nan
-    A_ch = np.zeros(6) * np.nan
-    A_ch_std = np.zeros(6) * np.nan
-
-    lc_height = 30.0e-2    # leaf chamber (ABT Sorime, fabrique en France), in m
-    lc_radius = 12.0e-2
-    skirt_height = np.array([11., 6., 11.5]) * 1e-2
-    # heights of plastic skirt at the bottom (m), approximated with paraboloid
-    n_leaves = np.array([87, 67,])  # leaf number in each branch chamber
-    A_leaf = 8.1 * 1e-4  # average area per leaf (m^2)
-    A_leaf_std = 1.5 * 1e-4 # standard deviation of leaf area (m^2)
-    sc_collar_height = np.array([5.40, 5.00, 9.45]) * 1e-2  # soil chamber collar height (m)
-
-    A_ch[0:2] = A_leaf * n_leaves  # leaf area in each branch chamber (m^2)
-    A_ch[2] = np.pi * lc_radius ** 2 + 2 * np.pi * lc_radius * lc_height  # a blank one, use inner surface area
-    A_ch[3:6] = 317.8 * 1e-4  # soil area covered by the chamber (m^2), from LI-8100A manual
-    A_ch_std[0:2] = A_leaf_std * n_leaves  # standard deviations of leaf area in each branch chamber (m^2)
-    A_ch_std[2:] = 0.  # for soil surface areas, no standard deviation is assigned
-    V_ch[0:3] = (lc_height + skirt_height * 0.5) * np.pi * lc_radius ** 2  # branch chamber volumes (m^3)
-    V_ch[3:6] = 4076.1 * 1e-6 + sc_collar_height * A_ch[3:6]  # soil chamber volume (m^3), bowl + collar volume
-
-    """
-    Chamber schedule
-    ----------------
-    For chamber #1
-    * Background calibration of analyzer: hh:00:00 - hh:02:00
-    * Atmospheric line: hh:02:00 - hh:03:00
-    * Chamber opening (before closure): hh:03:00 - hh:05:00
-    * Chamber closure: hh:05:00 - hh:13:00
-    * Chamber reopening (after closure): hh:13:00 - hh:15:00
-
-    Then add 15 minutes for the schedule of the next chamber measurements...
-    """
-    if doy < 105.5:  # from the beginning to 04/16/2013 12:00 (UTC)     
-        chlut = pd.DataFrame([1,2,3,4,5,6], columns=['ch_no'])
-        chlut['ch_label'] = ['LC1', 'LC2', 'LC3', 'SC1', 'SC2', 'SC3']
-        chlut['flowmeter_no'] = chlut['ch_no']
-        chlut['TC_no'] = chlut['ch_no']
-        chlut['PAR_no'] = [1, 2, 3, -1, -1, -1]
-        chlut['A_ch'] = A_ch[chlut['ch_no']-1]
-        chlut['A_ch_std'] = A_ch_std[chlut['ch_no']-1]
-        chlut['V_ch'] = V_ch[chlut['ch_no']-1]
-        chlut['ch_start'] = (2. + 15. * np.arange(6)) / 1440.  # minutes converted to day
-        chlut['ch_o_b'] = np.ones(6) / 1440. # with respect to ch_start
-        chlut['ch_cls'] = np.ones(6) * 3. / 1440. # with respect to ch_start
-        chlut['ch_o_a'] = np.ones(6) * 11. / 1440. # with respect to ch_start
-        chlut['ch_end'] = np.ones(6) * 13. / 1440. # with respect to ch_start
-        chlut['ch_atm_a'] = np.ones(6) * 13. / 1440. # with respect to ch_start
-        # if switch back to atmospheric line after chamber reopening, set the time with the key `chlut['ch_atm_a']`
-        # else, set the key to the same as `chlut['ch_end']`
-
-        smpl_cycle_len = 1.5 / 24.  # 1.5 hours per cycle
-        n_cycle_per_day = np.int(1. / smpl_cycle_len)  # 16 full cycles per day
-        n_ch = chlut['ch_no'].size
-        next_schedule_switch = 105.5
-
-    elif doy >= 105.5 and doy < 115.9:  # from 04/16/2013 12:00 to 04/26/2013 21:40 (UTC)
-        chlut = pd.DataFrame([1,2,4,1,2,5], columns=['ch_no'])
-        chlut['ch_label'] = ['LC1', 'LC2', 'SC1', 'LC1', 'LC2', 'SC2']
-        chlut['flowmeter_no'] = chlut['ch_no']
-        chlut['TC_no'] = chlut['ch_no']
-        chlut['PAR_no'] = [1, 2, -1, 1, 2, -1]
-        chlut['A_ch'] = A_ch[chlut['ch_no']-1]
-        chlut['A_ch_std'] = A_ch_std[chlut['ch_no']-1]
-        chlut['V_ch'] = V_ch[chlut['ch_no']-1]
-        chlut['ch_start'] = (2. + 15. * np.arange(6)) / 1440.  # minutes converted to day
-        chlut['ch_o_b'] = np.ones(6) / 1440. # with respect to ch_start
-        chlut['ch_cls'] = np.ones(6) * 3. / 1440. # with respect to ch_start
-        chlut['ch_o_a'] = np.ones(6) * 11. / 1440. # with respect to ch_start
-        chlut['ch_end'] = np.ones(6) * 13. / 1440. # with respect to ch_start
-        chlut['ch_atm_a'] = np.ones(6) * 13. / 1440. # with respect to ch_start
-        # if switch back to atmospheric line after chamber reopening, set the time with the key `chlut['ch_atm_a']`
-        # else, set the key to the same as `chlut['ch_end']`
-
-        smpl_cycle_len = 1.5 / 24.  # 1.5 hours per cycle
-        n_cycle_per_day = np.int(1. / smpl_cycle_len)  # 16 full cycles per day
-        n_ch = chlut['ch_no'].size
-        next_schedule_switch = 115.9
-
-    elif doy >= 115.9 and doy < 127. + 1./24.:  # from 04/26/2013 21:40 to 05/08/2013 01:00 (UTC)
-        chlut = pd.DataFrame([1,2,4,3,1,2,5,3], columns=['ch_no'])
-        chlut['ch_label'] = ['LC1', 'LC2', 'SC1', 'LC3', 'LC1', 'LC2', 'SC2', 'LC3']
-        chlut['flowmeter_no'] = chlut['ch_no']
-        chlut['TC_no'] = chlut['ch_no']
-        chlut['PAR_no'] = [1, 2, -1, 3, 1, 2, -1, 3]
-        chlut['A_ch'] = A_ch[chlut['ch_no']-1]
-        chlut['A_ch_std'] = A_ch_std[chlut['ch_no']-1]
-        chlut['V_ch'] = V_ch[chlut['ch_no']-1]
-        chlut['ch_start'] = (2. + 15. * np.arange(8)) / 1440.  # minutes converted to day
-        chlut['ch_o_b'] = np.ones(8) / 1440. # with respect to ch_start
-        chlut['ch_cls'] = np.ones(8) * 3. / 1440. # with respect to ch_start
-        chlut['ch_o_a'] = np.ones(8) * 11. / 1440. # with respect to ch_start
-        chlut['ch_end'] = np.ones(8) * 13. / 1440. # with respect to ch_start
-        chlut['ch_atm_a'] = np.ones(8) * 13. / 1440. # with respect to ch_start
-        # if switch back to atmospheric line after chamber reopening, set the time with the key `chlut['ch_atm_a']`
-        # else, set the key to the same as `chlut['ch_end']`
-
-        smpl_cycle_len = 2. / 24.  # 2 hours per cycle
-        n_cycle_per_day = np.int(1. / smpl_cycle_len)  # 12 full cycle per day
-        n_ch = chlut['ch_no'].size
-        next_schedule_switch = 127. + 1./24.
-
-    elif doy >= 127. + 1./24.:  # from 05/08/2013 01:00 (UTC) to the end
-        chlut = pd.DataFrame([1,2,4,6,1,2,5,6], columns=['ch_no'])
-        chlut['ch_label'] = ['LC1', 'LC2', 'SC1', 'SC3', 'LC1', 'LC2', 'SC2', 'SC3']
-        chlut['flowmeter_no'] = chlut['ch_no']
-        chlut['TC_no'] = chlut['ch_no']
-        chlut['PAR_no'] = [1, 2, -1, -1, 1, 2, -1, -1]
-        chlut['A_ch'] = A_ch[chlut['ch_no']-1]
-        chlut['A_ch_std'] = A_ch_std[chlut['ch_no']-1]
-        chlut['V_ch'] = V_ch[chlut['ch_no']-1]
-        chlut['ch_start'] = (2. + 15. * np.arange(8)) / 1440.  # minutes converted to day
-        chlut['ch_o_b'] = np.ones(8) / 1440. # with respect to ch_start
-        chlut['ch_cls'] = np.ones(8) * 3. / 1440. # with respect to ch_start
-        chlut['ch_o_a'] = np.ones(8) * 11. / 1440. # with respect to ch_start
-        chlut['ch_end'] = np.ones(8) * 13. / 1440. # with respect to ch_start
-        chlut['ch_atm_a'] = np.ones(8) * 13. / 1440. # with respect to ch_start
-        # if switch back to atmospheric line after chamber reopening, set the time with the key `chlut['ch_atm_a']`
-        # else, set the key to the same as `chlut['ch_end']`
-
-        smpl_cycle_len = 2. / 24.  # 2 hours per cycle
-        n_cycle_per_day = np.int(1. / smpl_cycle_len)  # 12 full cycle per day
-        n_ch = chlut['ch_no'].size
-        next_schedule_switch = np.nan
-    
-    if return_all:
-        return(chlut, n_ch, smpl_cycle_len, n_cycle_per_day, next_schedule_switch)
-    else:
-        return(chlut)
-
-def chamber_flow_rates(doy, ch_no):
-    """
-    Return chamber flow rates. 
-    
-    Use this function when flow rates are not in the biomet data tables. 
-    """
-    if ch_no <= 3:
-        flow_lpm = 1.5
-    if ch_no > 3:
-        flow_lpm = 2.
-    is_flow_STP = False
-    return(flow_lpm, is_flow_STP)
-
-def time_lag_opt_func():
-    return 0
-
-def volume_eff_opt_func():
-    return 0
-'''
-def conc_func(flow_rate, turnover_time, ):
-    return 0
-'''
 
 def conc_func(p, t):
     """
-    Calculate the concentration changes in chamber closure period as a function of time.
+    Calculate the changes in concentration in chamber closure period as a
+    function of time.
 
     Parameters
     ----------
-    p : parameter array with two elements
-        p[0]: fitted flux
-        p[1]: timelag / turnover time
-    t : time variable (normalized by the turnover time)
+    p : list or array with two elements
+        Parameter array
+            p[0]: fitted flux
+            p[1]: timelag / turnover time
+    t : array_like
+        Normalized time variable (by the turnover time).
 
     Returns
     -------
-    y : concentration changes in chamber closure period. 
+    y : array_like
+        Concentration changes in chamber closure period.
+
     """
     y = p[0] * (1. - np.exp(-t + p[1]))
-    return(y)
+    return y
+
 
 def resid_conc_func(p, t, y):
     """
-    Calculate the residuals of fitted concentration changes in chamber closure period as a function of time.
+    Calculate the residuals of fitted concentration changes in chamber closure
+    period as a function of time.
 
     Parameters
     ----------
-    p : parameter array with two elements
-        p[0]: fitted flux
-        p[1]: timelag / turnover time
-    t : time variable (normalized by the turnover time)
-    y : observations
+    p : list or array with two elements
+        Parameter array
+            p[0]: fitted flux
+            p[1]: timelag / turnover time
+    t : array_like
+        Normalized time variable (by the turnover time).
+    y : array_like
+        Observations of concentration changes in chamber closure period.
 
     Returns
     -------
-    resid : residuals of fitted concentration changes in chamber closure period. 
+    resid : array_like
+        Residuals of fitted concentration changes in chamber closure period.
+
     """
     resid = p[0] * (1. - np.exp(-t + p[1])) - y
-    return(resid)
+    return resid
 
-def IQR_func(x):
-    """ Calculate the interquartile range of an array. """
-    if np.sum(np.isfinite(x)) > 0:
-        q1, q3 = np.nanpercentile(x, [25,75])
-        return(q3 - q1)
+
+def resist_mean(x, IQR_range=1.5):
+    """
+    Calculate outlier-resistant mean of the sample using Tukey's outlier test.
+
+    Caveat: Does support calculation along an axis, unlike `numpy.mean()`.
+
+    Parameters
+    ----------
+    x : array_like
+        The sample.
+    IQR_range : float, optional
+        Parameter to control the inlier range defined by
+            [ Q_1 - IQR_range * (Q_3 - Q_1), Q_3 - IQR_range * (Q_3 - Q_1) ]
+        By default the parameter is 1.5, the original value used by John Tukey.
+
+    Returns
+    -------
+    x_rmean : float
+        The resistant mean of the sample with outliers removed.
+
+    References
+    ----------
+    .. [T77] John W. Tukey (1977). Exploratory Data Analysis. Addison-Wesley.
+
+    """
+    x = np.array(x)
+    if np.sum(np.isfinite(x)) <= 1:
+        return np.nanmean(x)
     else:
-        return(np.nan)
+        x_q1, x_q3 = np.nanpercentile(x, [25, 75])
+        x_iqr = x_q3 - x_q1
+        x_uplim = x_q3 + IQR_range * x_iqr
+        x_lolim = x_q1 - IQR_range * x_iqr
+        x_rmean = np.nanmean(x[(x >= x_lolim) & (x <= x_uplim)])
+        return x_rmean
 
 
+def resist_std(x, IQR_range=1.5):
+    """
+    Calculate outlier-resistant standard deviation of the sample using
+    Tukey's outlier test.
+
+    Caveat: Does support calculation along an axis, unlike `numpy.std()`.
+
+    Parameters
+    ----------
+    x : array_like
+        The sample.
+    IQR_range : float, optional
+        Parameter to control the inlier range defined by
+            [ Q_1 - IQR_range * (Q_3 - Q_1), Q_3 - IQR_range * (Q_3 - Q_1) ]
+        By default the parameter is 1.5, the original value used by John Tukey.
+
+    Returns
+    -------
+    x_rstd : float
+        The resistant standard deviation of the sample with outliers removed.
+        Degree of freedom = 1 is enforced for the sample standard deviation.
+
+    References
+    ----------
+    .. [T77] John W. Tukey (1977). Exploratory Data Analysis. Addison-Wesley.
+
+    """
+    x = np.array(x)
+    if np.sum(np.isfinite(x)) <= 1:
+        return(np.nanstd(x, ddof=1))
+    else:
+        x_q1, x_q3 = np.nanpercentile(x, [25, 75])
+        x_iqr = x_q3 - x_q1
+        x_uplim = x_q3 + IQR_range * x_iqr
+        x_lolim = x_q1 - IQR_range * x_iqr
+        x_rstd = np.nanstd(x[(x >= x_lolim) & (x <= x_uplim)], ddof=1)
+        return x_rstd
 
 
+def IQR_func(x, axis=None):
+    """
+    Calculate the interquartile range of an array.
+
+    Parameters
+    ----------
+    x : array_like
+        The sample.
+    axis : int, optional
+        Axis along which the percentiles are computed. Default is to ignore
+        and compute the flattened array.
+        (Same as the `axis` argument in `numpy.nanpercentile()`.)
+
+    Returns
+    -------
+    x_rstd : float or array_like
+        The resistant standard deviation of the sample with outliers removed.
+
+    """
+    if np.sum(np.isfinite(x)) > 0:
+        q1, q3 = np.nanpercentile(x, [25., 75.], axis=axis)
+        return q3 - q1
+    else:
+        return np.nan
 
 
 def p_sat_h2o(temp, ice=False, kelvin=False, method='gg'):
@@ -498,11 +385,6 @@ def dew_temp(e_sat, guess=25., kelvin=False, method='gg'):
     T_dew : float
         Dew temperature.
 
-    Raises
-    ------
-    AssertionError
-        If the argument `e_sat` is not float or int.
-
     Examples
     --------
     >>> dew_temp(3165)
@@ -516,10 +398,7 @@ def dew_temp(e_sat, guess=25., kelvin=False, method='gg'):
 
     """
     def __e_sat_residual(T, e_sat, kelvin, method):
-        return(p_sat_h2o(T, kelvin=kelvin, method=method) - e_sat)
-
-    assert type(e_sat) is float or type(e_sat) is int, \
-        "Takes a single value as the argument, array not allowed."
+        return p_sat_h2o(T, kelvin=kelvin, method=method) - e_sat
 
     if kelvin:
         guess += T_0